fix: change product structur table

+++ utils/validator.go
@@ -4,7 +4,7 @@ import "github.com/go-playground/validator/v10"
 
 type Product struct {
 	Name      string `validate:"required"`
-	Price     int    `validate:"required"`
+	Price     float64    `validate:"required"`
 	Stock     int    `validate:"required"`
 	Image_url string `validate:"required"`
 }
